Update signature of do_allocate() function.

+++ lib/custom_buffer.cc
@@ -97,7 +97,7 @@ void custom_buffer::post_work(int nitems)
     }
 }
 
-bool custom_buffer::do_allocate_buffer(int final_nitems, size_t sizeof_item)
+bool custom_buffer::do_allocate_buffer(size_t final_nitems, size_t sizeof_item)
 {
 #ifdef BUFFER_DEBUG
     {
